% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_districts_quebec.R
\docType{data}
\name{spatial_districts_quebec}
\alias{spatial_districts_quebec}
\title{Données géospatiales de la ville de Québec – Districts électoraux municipaux (2012)}
\format{
Un objet \code{sf} avec une colonne \code{geometry} représentant les polygones des circonscriptions.
}
\source{
Statistique Canada
}
\usage{
spatial_districts_quebec
}
\description{
Objet sf contenant les districts électoraux de la ville de Québec, provenant de \href{https://www.donneesquebec.ca/recherche/dataset/ca1847da-e908-43de-bd8c-588570331650}{Données Québec}.
}
\references{
VILLE DE QUÉBEC. Districts électoraux, \link{Jeu de données}, dans Données Québec, 2012, mis à jour le 24 février 2025. \link{https://www.donneesquebec.ca/recherche/dataset/vque_43}, (consulté le 05 mars 2025).
}
\keyword{datasets}
