% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_routes_quebec.R
\docType{data}
\name{spatial_routes_quebec}
\alias{spatial_routes_quebec}
\title{Données géospatiales de la ville de Québec – Routes}
\format{
Un objet \code{sf} avec une colonne \code{geometry} représentant les polygones des routes.
}
\source{
Données Québec
}
\description{
Objet sf contenant les districts électoraux de la ville de Québec, provenant de \href{https://www.donneesquebec.ca/recherche/dataset/ca1847da-e908-43de-bd8c-588570331650}{Données Québec}.
}
